Add files via upload
Problem Statement :--
 The objective is to analyze the flight
booking dataset obtained from a
platform which is used to book flight
tickets. A thorough study of the data
will aid in the discovery of valuable
insights that will be of enormous
value to passengers. Apply EDA,
statistical methods and Machine
learning algorithms in order to get
meaningful information from it. 


problem solving process :-
1> this project by importing all the necessary
libraries that will be required for the process.

2> Loading the data and removing unnecessary column from the dataframe.
3> Checking the shape of a dataframe and datatypes of
all columns along with calculating the statistical
data.
4> Checking out the missing values in a dataframe .
5> There is a variation in price with
different airlines. 
6> The price of the ticket
increases as the days left for
departure decreases 
7> Range of price of all the
flights of Economy and
Business class.
8> Range of price of flights with source
and destination city according to
the days left.
9> Visualization of categorical features with countplot .
10> Applying standardization and implementing Linear Regression , Decision Tree Regressor , Random Forest Regressor Model to
predict the price of a flight .